fix comment
